Make generated DTO encodeJSON() methods function

+++ kalix-example/src/main/java/eu/arrowhead/kalix/example/dto/Point.java
@@ -15,10 +15,10 @@ public interface Point {
      * @return X-coordinate.
      */
     @NameJSON("_x")
-    int x();
+    double x();
 
     /**
      * @return Y-coordinate, if any.
      */
-    int y();
+    Double y();
 }
